stm32n6,ExtMemLoader: build using CMake
The ExtMemLoader must have these symbols in the .elf so STM32CubeProgrammer can use it:
 - Init
 - Write
 - SectorErase
 - Verify
 - MassErase
 - StorageInfo

"noinline" and "used" attributes are required to ensure the symbols don't get optimized out. Same for "KEEP" in the linker script, to keep the section containing StorageInfo.

StorageInfo is in a special section which is not part of the program flashed in the STM32.

Also fix warning about different declaration of g_pfnVectors.